Move all run-related stuff into `run`
